services/backend/requirements.txt
fastapi==0.55.1
uvicorn==0.11.5
asyncpg==0.20.1
tortoise-orm==0.16.12
pytest==5.4.2
asynctest==0.13.0
requests==2.23.0
gunicorn==20.0.4
pytest-cov==2.9.0
flake8===3.8.2
black==19.10b0
isort==4.3.21